Split oversized Rust files to satisfy 2000-line lint gate
- codegen.rs (2723) → codegen/mod.rs (574) + render_expr.rs (560) + render_stmt.rs (841)
- rumoca-phase-dae/lib.rs (2469) → lib.rs (621) + variable_analysis.rs (1166) + dae_lowering.rs (746)
- functions.rs (2120) → functions.rs (1443) + function_lowering.rs (693)
