+++ crates/rumoca-phase-codegen/src/codegen/codegen_tests.rs
@@ -150,8 +150,7 @@ fn test_render_array_comprehension_expression() {
 
     let rendered =
         render_expression(&Value::from_serialize(&expr), &ExprConfig::default()).unwrap();
-    // In Python mode (default), uses list comprehension with range()
-    assert_eq!(rendered, "[i for i in range(1, (n) + 1) if (i > 0)]");
+    assert_eq!(rendered, "{i for i in 1:n if (i > 0)}");
 }
 
 #[test]
@@ -405,425 +404,5 @@ fn test_template_missing_assignment_target_fails_fast() {
         msg.contains("Assignment missing 'comp' field")
             || msg.contains("target resolved to empty component reference"),
         "expected strict assignment error, got: {msg}"
-    );
-}
-
-#[test]
-fn test_fmi2_model_description_empty_dae() {
-    let dae = dae::Dae::new();
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL_DESCRIPTION, "TestModel")
-            .unwrap();
-    assert!(result.contains("fmiVersion=\"2.0\""));
-    assert!(result.contains("modelName=\"TestModel\""));
-    assert!(result.contains("guid=\"TestModel-rumoca\""));
-    assert!(result.contains("<ModelExchange modelIdentifier=\"TestModel\""));
-    assert!(result.contains("<ModelVariables>"));
-    assert!(result.contains("</ModelVariables>"));
-    assert!(result.contains("<ModelStructure>"));
-    assert!(result.contains("generationTool=\"Rumoca\""));
-}
-
-#[test]
-fn test_fmi2_model_description_with_states_and_params() {
-    let mut dae = dae::Dae::new();
-    dae.states.insert(
-        "x".into(),
-        rumoca_ir_dae::Variable {
-            name: "x".into(),
-            start: Some(rumoca_ir_dae::Expression::Literal(
-                rumoca_ir_dae::Literal::Real(1.0),
-            )),
-            unit: Some("m".to_string()),
-            description: Some("position".to_string()),
-            ..Default::default()
-        },
-    );
-    dae.states.insert(
-        "v".into(),
-        rumoca_ir_dae::Variable {
-            name: "v".into(),
-            start: Some(rumoca_ir_dae::Expression::Literal(
-                rumoca_ir_dae::Literal::Real(0.0),
-            )),
-            unit: Some("m/s".to_string()),
-            ..Default::default()
-        },
-    );
-    dae.parameters.insert(
-        "k".into(),
-        rumoca_ir_dae::Variable {
-            name: "k".into(),
-            start: Some(rumoca_ir_dae::Expression::Literal(
-                rumoca_ir_dae::Literal::Real(10.0),
-            )),
-            ..Default::default()
-        },
-    );
-
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL_DESCRIPTION, "SpringMass")
-            .unwrap();
-
-    // States appear as local/continuous with exact initial
-    assert!(result.contains("name=\"x\""));
-    assert!(result.contains("causality=\"local\""));
-    assert!(result.contains("variability=\"continuous\""));
-    assert!(result.contains("initial=\"exact\""));
-    assert!(result.contains("start=\"1.0\""));
-    assert!(result.contains("unit=\"m\""));
-    assert!(result.contains("description=\"position\""));
-
-    // Derivatives appear
-    assert!(result.contains("name=\"der(x)\""));
-    assert!(result.contains("name=\"der(v)\""));
-
-    // Parameters appear as parameter/fixed
-    assert!(result.contains("name=\"k\""));
-    assert!(result.contains("causality=\"parameter\""));
-    assert!(result.contains("variability=\"fixed\""));
-    assert!(result.contains("start=\"10.0\""));
-
-    // Model structure sections
-    assert!(result.contains("<Derivatives>"));
-    assert!(result.contains("<InitialUnknowns>"));
-}
-
-#[test]
-fn test_fmi2_model_c_empty_dae() {
-    let dae = dae::Dae::new();
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL, "TestModel").unwrap();
-    assert!(result.contains("FMI 2.0 Model Exchange"));
-    assert!(result.contains("#define N_STATES"));
-    assert!(result.contains("fmi2Instantiate"));
-    assert!(result.contains("fmi2GetDerivatives"));
-    assert!(result.contains("fmi2SetContinuousStates"));
-    assert!(result.contains("fmi2GetReal"));
-    assert!(result.contains("fmi2SetReal"));
-    assert!(result.contains("fmi2Terminate"));
-    assert!(result.contains("MODEL_GUID"));
-}
-
-#[test]
-fn test_fmi2_model_c_with_states_and_params() {
-    let mut dae = dae::Dae::new();
-    dae.states.insert(
-        "x".into(),
-        rumoca_ir_dae::Variable {
-            name: "x".into(),
-            start: Some(rumoca_ir_dae::Expression::Literal(
-                rumoca_ir_dae::Literal::Real(1.0),
-            )),
-            nominal: Some(rumoca_ir_dae::Expression::Literal(
-                rumoca_ir_dae::Literal::Real(5.0),
-            )),
-            ..Default::default()
-        },
-    );
-    dae.parameters.insert(
-        "k".into(),
-        rumoca_ir_dae::Variable {
-            name: "k".into(),
-            start: Some(rumoca_ir_dae::Expression::Literal(
-                rumoca_ir_dae::Literal::Real(10.0),
-            )),
-            ..Default::default()
-        },
-    );
-
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL, "SpringMass").unwrap();
-
-    // Dimensions
-    assert!(result.contains("#define N_STATES         1"));
-    assert!(result.contains("#define N_PARAMETERS     1"));
-
-    // Default initialization (two-phase: alias = expr; m->p[i] = alias)
-    assert!(result.contains("m->x[0] = 1.0;"));
-    assert!(result.contains("k = 10.0;"));
-    assert!(result.contains("m->p[0] = k;"));
-
-    // Nominals
-    assert!(result.contains("x_nominal[0] = 5.0;"));
-
-    // VR offsets
-    assert!(result.contains("#define VR_X      0"));
-    assert!(result.contains("#define VR_XDOT   N_STATES"));
-}
-
-#[test]
-fn test_fmi2_templates_value_reference_consistency() {
-    // Ensure both templates produce consistent value reference layouts
-    let mut dae = dae::Dae::new();
-    dae.states.insert(
-        "x".into(),
-        rumoca_ir_dae::Variable {
-            name: "x".into(),
-            ..Default::default()
-        },
-    );
-    dae.inputs.insert(
-        "u".into(),
-        rumoca_ir_dae::Variable {
-            name: "u".into(),
-            ..Default::default()
-        },
-    );
-    dae.outputs.insert(
-        "w".into(),
-        rumoca_ir_dae::Variable {
-            name: "w".into(),
-            ..Default::default()
-        },
-    );
-
-    let xml =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL_DESCRIPTION, "M").unwrap();
-    let c_code = render_template_with_name(&dae, crate::templates::FMI2_MODEL, "M").unwrap();
-
-    // x at vr=0, xdot at vr=1 (n_x=1), u at vr=2 (n_y=0), w at vr=3
-    assert!(result_contains_vr(&xml, "x", 0));
-    assert!(result_contains_vr(&xml, "der(x)", 1));
-    assert!(result_contains_vr(&xml, "u", 2));
-    assert!(result_contains_vr(&xml, "w", 3));
-
-    // C code should have matching offsets
-    assert!(c_code.contains("#define VR_X      0"));
-}
-
-fn result_contains_vr(xml: &str, var_name: &str, vr: usize) -> bool {
-    let pattern = format!("name=\"{}\" valueReference=\"{}\"", var_name, vr);
-    xml.contains(&pattern)
-}
-
-#[test]
-fn test_fmi2_model_c_has_cosimulation_support() {
-    let dae = dae::Dae::new();
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL, "TestModel").unwrap();
-
-    // Co-Simulation API functions
-    assert!(result.contains("fmi2DoStep"));
-    assert!(result.contains("fmi2CancelStep"));
-    assert!(result.contains("forward Euler"));
-
-    // Should accept both ME and CS
-    assert!(result.contains("fmuType != fmi2ModelExchange && fmuType != fmi2CoSimulation"));
-}
-
-#[test]
-fn test_fmi2_model_description_has_cosimulation() {
-    let dae = dae::Dae::new();
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL_DESCRIPTION, "TestModel")
-            .unwrap();
-
-    assert!(result.contains("<CoSimulation"));
-    assert!(result.contains("<ModelExchange"));
-    assert!(result.contains("canHandleVariableCommunicationStepSize"));
-}
-
-#[test]
-fn test_fmi2_model_description_derivative_indices_are_contiguous_for_mixed_arrays() {
-    let mut dae = dae::Dae::new();
-    dae.states.insert(
-        "x".into(),
-        rumoca_ir_dae::Variable {
-            name: "x".into(),
-            dims: vec![2],
-            ..Default::default()
-        },
-    );
-    dae.states.insert(
-        "v".into(),
-        rumoca_ir_dae::Variable {
-            name: "v".into(),
-            dims: vec![3],
-            ..Default::default()
-        },
-    );
-
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL_DESCRIPTION, "IdxModel")
-            .unwrap();
-
-    // n_x = 5, derivative block must be indices 6..10 (1-based).
-    for idx in 6..=10 {
-        assert!(
-            result.contains(&format!("<Unknown index=\"{idx}\"/>")),
-            "missing derivative index {idx} in rendered XML"
-        );
-    }
-}
-
-#[test]
-fn test_fmi2_model_c_has_output_and_discrete_functions() {
-    let mut dae = dae::Dae::new();
-    dae.outputs.insert(
-        "w".into(),
-        rumoca_ir_dae::Variable {
-            name: "w".into(),
-            ..Default::default()
-        },
-    );
-
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL, "TestModel").unwrap();
-
-    // Output computation
-    assert!(result.contains("compute_outputs"));
-    // Discrete update computation
-    assert!(result.contains("compute_discrete_updates"));
-    // Called from fmi2NewDiscreteStates
-    assert!(result.contains("compute_discrete_updates(m)"));
-    // Outputs should no longer be hardcoded TODO stubs.
-    assert!(!result.contains("TODO: output equation"));
-}
-
-#[test]
-fn test_fmi2_model_c_maps_array_state_derivatives_by_indexed_name() {
-    use rumoca_ir_core::OpBinary;
-    use rumoca_ir_dae::{BuiltinFunction, Equation, Expression, Literal};
-
-    let mut dae = dae::Dae::new();
-    dae.states.insert(
-        "x".into(),
-        rumoca_ir_dae::Variable {
-            name: "x".into(),
-            dims: vec![2],
-            ..Default::default()
-        },
-    );
-
-    // 0 = der(x[1]) - 1
-    dae.f_x.push(Equation::residual(
-        Expression::Binary {
-            op: OpBinary::Sub(Default::default()),
-            lhs: Box::new(Expression::BuiltinCall {
-                function: BuiltinFunction::Der,
-                args: vec![Expression::VarRef {
-                    name: "x[1]".into(),
-                    subscripts: vec![],
-                }],
-            }),
-            rhs: Box::new(Expression::Literal(Literal::Real(1.0))),
-        },
-        Default::default(),
-        "test",
-    ));
-    // 0 = der(x[2]) - 2
-    dae.f_x.push(Equation::residual(
-        Expression::Binary {
-            op: OpBinary::Sub(Default::default()),
-            lhs: Box::new(Expression::BuiltinCall {
-                function: BuiltinFunction::Der,
-                args: vec![Expression::VarRef {
-                    name: "x[2]".into(),
-                    subscripts: vec![],
-                }],
-            }),
-            rhs: Box::new(Expression::Literal(Literal::Real(2.0))),
-        },
-        Default::default(),
-        "test",
-    ));
-
-    let result =
-        render_template_with_name(&dae, crate::templates::FMI2_MODEL, "ArrayDeriv").unwrap();
-    assert!(result.contains("m->xdot[0] = 1.0;"));
-    assert!(result.contains("m->xdot[1] = 2.0;"));
-}
-
-#[test]
-fn test_fmi2_model_c_exposes_discrete_valued_via_get_real() {
-    let mut dae = dae::Dae::new();
-    dae.discrete_valued.insert(
-        "mode".into(),
-        rumoca_ir_dae::Variable {
-            name: "mode".into(),
-            ..Default::default()
-        },
-    );
-
-    let result = render_template_with_name(&dae, crate::templates::FMI2_MODEL, "DiscVal").unwrap();
-    assert!(result.contains("vr >= VR_M && vr < VR_M + N_DISCRETE_VAL"));
-}
-
-#[test]
-fn test_is_self_call_detects_self_referential_function() {
-    use rumoca_ir_dae::{
-        ComponentRefPart, ComponentReference, Dae, Expression, Function, FunctionParam, Statement,
-        VarName,
-    };
-
-    fn comp_ref(name: &str) -> ComponentReference {
-        ComponentReference {
-            local: false,
-            parts: vec![ComponentRefPart {
-                ident: name.to_string(),
-                subs: vec![],
-            }],
-            def_id: None,
-        }
-    }
-
-    fn func_param(name: &str) -> FunctionParam {
-        FunctionParam::new(name, "Real")
-    }
-
-    // Create a function like Modelica.Math.sin whose body calls itself
-    let mut dae = Dae::new();
-    let func = Function {
-        name: VarName("Modelica.Math.sin".to_string()),
-        inputs: vec![func_param("u")],
-        outputs: vec![func_param("y")],
-        body: vec![Statement::Assignment {
-            comp: comp_ref("y"),
-            value: Expression::FunctionCall {
-                name: VarName("Modelica.Math.sin".to_string()),
-                args: vec![Expression::VarRef {
-                    name: VarName("u".to_string()),
-                    subscripts: vec![],
-                }],
-                is_constructor: false,
-            },
-        }],
-        ..Function::new("Modelica.Math.sin", Default::default())
-    };
-    dae.functions
-        .insert(VarName("Modelica.Math.sin".to_string()), func);
-
-    // Render a template that uses is_self_call
-    let template = r#"{% for func_name, func in dae.functions | items %}{{ is_self_call(func_name, func) }}{% endfor %}"#;
-    let result = render_template(&dae, template).unwrap();
-    assert_eq!(
-        result.trim(),
-        "true",
-        "is_self_call should detect self-referential function"
-    );
-
-    // Also test with a non-self-referential function
-    let mut dae2 = Dae::new();
-    let func2 = Function {
-        name: VarName("myFunc".to_string()),
-        inputs: vec![func_param("u")],
-        outputs: vec![func_param("y")],
-        body: vec![Statement::Assignment {
-            comp: comp_ref("y"),
-            value: Expression::VarRef {
-                name: VarName("u".to_string()),
-                subscripts: vec![],
-            },
-        }],
-        ..Function::new("myFunc", Default::default())
-    };
-    dae2.functions.insert(VarName("myFunc".to_string()), func2);
-
-    let result2 = render_template(&dae2, template).unwrap();
-    assert_eq!(
-        result2.trim(),
-        "false",
-        "is_self_call should not detect non-self-referential function"
     );
 }
